refactor exercise creator
