+++ exerciser/package.json
@@ -10,12 +10,12 @@
     "src"
   ],
   "dependencies": {
-    "@theia/core": "1.55.0",
-    "@theia/ai-core": "1.55.0",
-    "@theia/filesystem": "1.55.0",
     "@theia/ai-chat": "1.55.0",
     "@theia/ai-chat-ui": "1.55.0",
-    "@theia/ai-terminal": "1.55.0"
+    "@theia/ai-core": "1.55.0",
+    "@theia/ai-terminal": "1.55.0",
+    "@theia/core": "1.55.0",
+    "@theia/filesystem": "1.55.0"
   },
   "devDependencies": {
     "rimraf": "^5.0.0",
@@ -30,7 +30,6 @@
   "theiaExtensions": [
     {
       "frontend": "lib/browser/exerciser-frontend-module"
-
     }
   ]
 }
