+++ exerciser/src/browser/exercise-service/types.ts
@@ -1,0 +1,13 @@
+export interface ExerciseFile{
+    filename: string;
+    content: string;
+}
+
+export  interface Exercise {
+    exerciseId:string;
+    exerciseName: string;
+    exerciseSummarization: string;
+    fileListSummarization: string;
+    exerciseFiles: ExerciseFile[];
+    conductorFiles: ExerciseFile[];
+}
